requirements.txt
requests==2.25.1
python-magic==0.4.27